Change minimal visibility regex
Currently when the following metar is parsed `KATW 022045Z 0000KT 10SM SCT120 00/M08 A2996`, we get an error `UnexpectedParseError: container.visibility not instantiated`. This happens because wind string (`0000KT`) is matching minimal visibility regex. Suggest to update the minimal visibility to more string, so it does not intersect with wind string.